refactor: theme, implement `getLayout`
